Add defenderCaseFileCategoriesForRestrictionAndInvestigationCases to limitedAccessCase.service.ts
