+++ apps/judicial-system/backend/src/app/modules/file/index.ts
@@ -1,2 +1,3 @@
 export { CaseFile } from './models/file.model'
 export { FileService } from './file.service'
+export { defenderCaseFileCategoriesForRestrictionAndInvestigationCases } from './guards/caseFileCategory'
