+++ libs/judicial-system/types/src/lib/notification.ts
@@ -15,16 +15,3 @@ export enum NotificationType {
   APPEAL_JUDGES_ASSIGNED = 'APPEAL_JUDGES_ASSIGNED',
   APPEAL_CASE_FILES_UPDATED = 'APPEAL_CASE_FILES_UPDATED',
 }
-
-export interface Recipient {
-  success: boolean
-  address?: string
-}
-
-export interface Notification {
-  id: string
-  created: string
-  caseId: string
-  type: NotificationType
-  recipients: Recipient[]
-}
